+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * issue: ReportIt was not properly parsing maxRRDValues of |query_ifSpeed| and |query_ifHighSpeed|
 * feature: Dropping support for Cacti 1.2.x
 * feature: Remove archive and export functionality.  Using Cacti's report log instead
+* feature: Add Notification Lists to Email Settings
+* feature: Add string based email and bcc to Email Settings
 
 --- 2.0.0 ---